+++ tests/templates/kuttl/oidc/login.py
@@ -10,9 +10,10 @@ logging.basicConfig(
 )
 
 session = requests.Session()
+url = "http://airflow-webserver-default:8080"
 
 # Click on "Sign In with keycloak" in Airflow
-login_page = session.get("http://airflow-webserver:8080/login/keycloak?next=")
+login_page = session.get(f"{url}/login/keycloak?next=")
 
 assert login_page.ok, "Redirection from Airflow to Keycloak failed"
 assert login_page.url.startswith(
@@ -27,32 +28,32 @@ welcome_page = session.post(
 )
 
 assert welcome_page.ok, "Login failed"
-assert (
-    welcome_page.url == "http://airflow-webserver:8080/home"
-), "Redirection to the Airflow home page expected"
+assert welcome_page.url == f"{url}/home", (
+    "Redirection to the Airflow home page expected"
+)
 
 # Open the user information page in Airflow
-userinfo_page = session.get("http://airflow-webserver:8080/users/userinfo/")
+userinfo_page = session.get(f"{url}/users/userinfo/")
 
 assert userinfo_page.ok, "Retrieving user information failed"
-assert (
-    userinfo_page.url == "http://airflow-webserver:8080/users/userinfo/"
-), "Redirection to the Airflow user info page expected"
+assert userinfo_page.url == f"{url}/users/userinfo/", (
+    "Redirection to the Airflow user info page expected"
+)
 
 # Expect the user data provided by Keycloak in Airflow
 userinfo_page_html = BeautifulSoup(userinfo_page.text, "html.parser")
 table_rows = userinfo_page_html.find_all("tr")
 user_data = {tr.find("th").text: tr.find("td").text for tr in table_rows}
 
-assert (
-    user_data["First Name"] == "Jane"
-), "The first name of the user in Airflow should match the one provided by Keycloak"
-assert (
-    user_data["Last Name"] == "Doe"
-), "The last name of the user in Airflow should match the one provided by Keycloak"
-assert (
-    user_data["Email"] == "jane.doe@stackable.tech"
-), "The email of the user in Airflow should match the one provided by Keycloak"
+assert user_data["First Name"] == "Jane", (
+    "The first name of the user in Airflow should match the one provided by Keycloak"
+)
+assert user_data["Last Name"] == "Doe", (
+    "The last name of the user in Airflow should match the one provided by Keycloak"
+)
+assert user_data["Email"] == "jane.doe@stackable.tech", (
+    "The email of the user in Airflow should match the one provided by Keycloak"
+)
 
 # Later this can be extended to use different OIDC providers (currently only Keycloak is
 # supported)